requirements.txt
streamlit==1.45.1
requests==2.32.3
plotly==6.0.1
python-dotenv==1.0.1
pandas==2.2.3
